Fix response.body property access for StreamingResponse (#124)
In my versioned API route, the response class is a `StreamingResponse`. 
When forming the `ResponseInfo`, I got an `AttributeError`. 

I found that there was an assumption that all Starlette Responses have a
`.body` attribute, but `StreamingResponse` doesn't(`body_iterator`
instead). Therefore, I suggest this change to check if the response
object has a body attribute instead of using `isinstance()`.

P.S.
Overall, I started using this library and liked its concept of
migrations, because full code copying was a nightmare without any light
at the end of the tunnel.

Looking forward to your response and comments!

---------

Co-authored-by: Stanislav Zmiev <zmievsa@gmail.com>