Update `actions/(upload|download)-artifact` to v4
`actions/upload-artifact` v3 and `actions/download-artifact` v3
will be disabled on 2025-01-31.

https://github.blog/changelog/2024-04-16-deprecation-notice-v3-of-the-artifact-actions/

The only change from v3 to v4 that affects this workflow
is that artifacts with the same name can only be overwritten,
so the artifact names must now be unique.

Merging now happens at download, not at upload,
so this commit modifies the Python and docs test artifact names
to share the same prefix and downloads them all in one shot.