Copy field annotations when building versioned models with future annotations (#257)
Closes #256.

In the _specific_ case of WithJsonSchema this may have in-intended
consequences for the schema versioning (where changes in the override
via the
annotation aren't detected), but I think generally this is the correct
behaviour.

The `_rebuild_annotated` returns a value that is the same as if future
annotations wasn't in scope.